got home and all podcass working with the new routing
