+++ client/src/containers/HomeContainer.jsx
@@ -11,7 +11,7 @@ import Header from '../components/Header'
 import Carousel from "../components/Carousel"
 
 
-const HomeContainer = ({podcasts, loading}) => {
+const HomeContainer = ({podcasts}) => {
 
   return (
     <>
