+++ client/src/services/PodcastsServices.jsx
@@ -1,0 +1,6 @@
+const podcastsURL = "http://localhost:9000/api/podcasts"
+
+export const getAllPodcasts = () => {
+  return fetch(podcastsURL)
+    .then(response => response.json())
+}
